Move FFI Array class to separate file
